refactor agent
