+++ client-app/src/features/activities/form/ActivityForm.tsx
@@ -59,6 +59,7 @@ export default function ActivityForm({
         />
         <Form.Input
           placeholder="Date"
+          type="date"
           value={activity.date}
           name="date"
           onChange={handleInputChange}
