+++ client-app/src/features/details/ActivityDetails.tsx
